refactor: colonizethis_economy phase 8 splits, 260-line ratchet, test mirror
Split near-cap economy lib modules (FRR credits, purchased-tile riches,
trade counsel need analysis, trade interception odds), lower the
economy source-file ceiling 310→260, relocate flat-root tests under
test/economy/, and add repo.economy_test_mirrors_lib enforcement.

Refs #4299
