+++ SPEC/program/repo-lint.md
@@ -241,7 +241,8 @@
 | `tool/check_combat_test_support_loc.dart` | Ratchet ceiling for `packages/colonizethis_combat_test_support/lib/**` physical Dart LOC (#4196 ≤7250, measured ≈7242 post-split; baseline was 7,307); rule `repo.combat_test_support_loc` (Refs #4196) |
 | `tool/check_combat_test_support_file_size.dart` | Ratchet ceiling for `packages/colonizethis_combat_test_support/lib/**` physical lines per file (#4196 ≤220, chosen just above the post-split largest file `quick_battle_input_builder_scenarios.dart` = 220); generated suffixes excluded; rule `repo.combat_test_support_file_size` (Refs #4196) |
 | `tool/check_combat_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_combat/lib/**` physical lines per file (phase-4 ≤280, chosen just above the post-split largest file `quick_battle_resolver_engine.dart` ≈268; stricter than the shared 550-line `repo.domain_package_source_file_size` cap); generated suffixes excluded; shrink-only grandfather allowlist during transition; rule `repo.combat_lib_file_size` (Refs #4284) |
-| `tool/check_economy_source_file_size.dart` | Ratchet ceiling for `packages/colonizethis_economy/lib/**` physical lines per file (phase-7 ≤310, chosen just above the post-split largest file `world_market/first_right_credits.dart` ≈302; stricter than the shared 550-line `repo.domain_package_source_file_size` cap) so the #4049 kitchen-sink splits (`treasury_bid_budget.dart` → spend/available/caps siblings, `town_manufacturing_bonus.dart` → math/delivered-raw siblings, `first_right_credits.dart` → embassy-kickback sibling, `tile_extraction_pipeline.dart` → context sibling) cannot silently re-grow; generated suffixes excluded; rule `repo.economy_source_file_size` (Refs #4049) |
+| `tool/check_economy_source_file_size.dart` | Ratchet ceiling for `packages/colonizethis_economy/lib/**` physical lines per file (phase-8 ≤260 after #4299 near-cap splits; phase-7 was ≤310) so kitchen-sink modules cannot silently re-grow; stricter than the shared 550-line `repo.domain_package_source_file_size` cap; generated suffixes excluded; rule `repo.economy_source_file_size` (Refs #4049, #4299) |
+| `tool/check_economy_test_mirrors_lib.dart` | Keep the `packages/colonizethis_economy/test/**` tree mirrored to `lib/src/economy/` per the testing rule (`test/` mirrors `lib/`): any Dart source that sits **directly** at the `test/` root is a violation — both `*_test.dart` files and co-located support helpers must live in a subdirectory such as `test/economy/` or `test/economy/world_market/` that mirrors `lib/src/economy/`. Complements the testing-rule policy (Refs #4299); rule `repo.economy_test_mirrors_lib` |
 | `tool/check_save_source_file_size.dart` | PR-blocking physical-line gate: every `.dart` file under `packages/colonizethis_save/lib/src/**` must stay at or below **400** physical lines (empty shrink-only allowlist); keeps the #4077 `GameSaveAdapter` concern splits from re-merging; rule `repo.save_source_file_size` (Refs #4077) |
 | `tool/check_save_test_file_size.dart` | PR-blocking physical-line gate: every `.dart` file under `packages/colonizethis_save/test/**` must stay at or below **400** physical lines (empty shrink-only allowlist; mirrors `repo.economy_test_file_size` / `repo.logic_test_file_size`); rule `repo.save_test_file_size` (Refs #4077) |
 | `tool/check_economy_test_no_local_run_scenarios.dart` | Forbid local `void Function() run` scenario shells under `packages/colonizethis_economy/test/**`; pins live in `colonizethis_economy_test_support` (Refs #4014); rule `repo.economy_test_no_local_run_scenarios` |
