+++ packages/colonizethis_economy/lib/src/economy/trade_counsel/trade_counsel_speculative_bids.dart
@@ -1,0 +1,67 @@
+/// Speculative bid need selection for trade counsel (neutral treasury path).
+library;
+
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import 'trade_counsel_constants.dart';
+
+void tradeCounselAddSpeculativeBidNeeds({
+  required Map<CommodityId, int> need,
+  required Map<CommodityId, int> available,
+  required Stockpile projected,
+  required Map<CommodityId, int> carryForwardBids,
+  required WorldMarketState state,
+}) {
+  bool eligible(CommodityId id) {
+    if (richesCommodityIds.contains(id)) return false;
+    if (need.containsKey(id)) return false;
+    if (available.containsKey(id)) return false;
+    final projectedQty = projected.quantityOf(id);
+    final carryQty = carryForwardBids[id] ?? 0;
+    return kTradeCounselSpeculativeBidStockpileTarget -
+            projectedQty -
+            carryQty >
+        0;
+  }
+
+  int gapFor(CommodityId id) {
+    final projectedQty = projected.quantityOf(id);
+    final carryQty = carryForwardBids[id] ?? 0;
+    return kTradeCounselSpeculativeBidStockpileTarget - projectedQty - carryQty;
+  }
+
+  int offerVolumeFor(CommodityId id) =>
+      state.lastTurnActivity[id]?.totalOfferQuantity ?? 0;
+
+  final eligibleIds = CommodityCatalog.all
+      .map((c) => c.id)
+      .where(eligible)
+      .toList(growable: false);
+  if (eligibleIds.isEmpty) return;
+
+  CommodityId pick;
+  final liquid = eligibleIds.where((id) => offerVolumeFor(id) > 0).toList()
+    ..sort((a, b) {
+      final volCmp = offerVolumeFor(b).compareTo(offerVolumeFor(a));
+      if (volCmp != 0) return volCmp;
+      return a.compareTo(b);
+    });
+  if (liquid.isNotEmpty) {
+    pick = liquid.first;
+  } else {
+    final foods = eligibleIds
+        .where(
+          (id) => CommodityCatalog.byId[id]?.category == CommodityCategory.food,
+        )
+        .toList()
+      ..sort();
+    if (foods.isNotEmpty) {
+      pick = foods.first;
+    } else {
+      final sortedEligible = [...eligibleIds]..sort();
+      pick = sortedEligible.first;
+    }
+  }
+  need[pick] = gapFor(pick);
+}
