+++ packages/colonizethis_economy/lib/src/economy/trade_counsel/trade_counsel_stockpile_projection.dart
@@ -1,0 +1,87 @@
+/// Stockpile projection and input-need helpers for trade counsel.
+library;
+
+import 'package:colonizethis_data/colonizethis_data.dart';
+import 'package:colonizethis_models/colonizethis_models.dart';
+
+import '../industry_counsel/industry_counsel_constants.dart'
+    show kIndustryCounselShortageThreshold;
+
+Stockpile tradeCounselProjectStockpileAfterProduction({
+  required Stockpile stockpile,
+  required List<AssignedRecipe> productionAssignments,
+}) {
+  var projected = stockpile;
+  for (final assignment in productionAssignments) {
+    final recipe = ProductionRecipesCatalog.byId[assignment.recipeId];
+    if (recipe == null || assignment.assignedLabour <= 0) continue;
+    final runs = assignment.assignedLabour ~/ recipe.labourPerOutput;
+    if (runs <= 0) continue;
+    for (final entry in recipe.inputQuantities.entries) {
+      projected = projected.applyDelta(entry.key, -entry.value * runs);
+    }
+    projected = projected.applyDelta(
+      recipe.outputCommodityId,
+      recipe.outputQuantity * runs,
+    );
+  }
+  return projected;
+}
+
+Set<CommodityId> tradeCounselTrackedCommodityIds({
+  required Stockpile stockpile,
+  required Stockpile projected,
+  required Map<CommodityId, int> inputNeeds,
+  required List<AssignedRecipe> productionAssignments,
+}) {
+  final ids = <CommodityId>{...inputNeeds.keys};
+  for (final entry in stockpile.quantities.entries) {
+    if (entry.value > 0) ids.add(entry.key);
+  }
+  for (final entry in projected.quantities.entries) {
+    if (entry.value > 0) ids.add(entry.key);
+  }
+  for (final assignment in productionAssignments) {
+    final recipe = ProductionRecipesCatalog.byId[assignment.recipeId];
+    if (recipe == null) continue;
+    ids.add(recipe.outputCommodityId);
+    ids.addAll(recipe.inputQuantities.keys);
+  }
+  for (final commodity in CommodityCatalog.all) {
+    if (commodity.category == CommodityCategory.food) {
+      ids.add(commodity.id);
+    }
+  }
+  return ids;
+}
+
+int tradeCounselConsumptionForecast({
+  required CommodityId commodityId,
+  required Commodity commodity,
+  required Map<CommodityId, int> inputNeeds,
+}) {
+  if (inputNeeds.containsKey(commodityId)) {
+    return inputNeeds[commodityId]!.clamp(1, kIndustryCounselShortageThreshold);
+  }
+  if (commodity.category == CommodityCategory.food) {
+    return kIndustryCounselShortageThreshold;
+  }
+  return (kIndustryCounselShortageThreshold ~/ 2)
+      .clamp(1, kIndustryCounselShortageThreshold);
+}
+
+Map<CommodityId, int> tradeCounselInputNeedsFromAssignments(
+  List<AssignedRecipe> productionAssignments,
+) {
+  final needs = <CommodityId, int>{};
+  for (final assignment in productionAssignments) {
+    final recipe = ProductionRecipesCatalog.byId[assignment.recipeId];
+    if (recipe == null || assignment.assignedLabour <= 0) continue;
+    final runs = assignment.assignedLabour ~/ recipe.labourPerOutput;
+    if (runs <= 0) continue;
+    for (final entry in recipe.inputQuantities.entries) {
+      needs[entry.key] = (needs[entry.key] ?? 0) + entry.value * runs;
+    }
+  }
+  return needs;
+}
