+++ packages/colonizethis_economy/lib/src/economy/trade_interception.dart
@@ -5,8 +5,10 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_world/colonizethis_world.dart';
 
 import 'commodity_totals.dart';
-import 'trade_interception_constants.dart';
+import 'trade_interception_odds.dart';
 import 'trade_interception_scan.dart';
+
+export 'trade_interception_odds.dart';
 
 /// Trade/transport interception during extraction auto-transport.
 /// SPEC/program/naval-movement-resolution.md (P_cargo_intercept, P_ship_sunk);
@@ -17,7 +19,7 @@ import 'trade_interception_scan.dart';
 /// testable concerns (Refs #3427 step 7). Tuning constants live in
 /// `trade_interception_constants.dart` and the fleet aggregation in
 /// `trade_interception_scan.dart` (Refs #3615 Cluster 4 file decomposition).
-/// Odds / cargo / ship-loss helpers extracted Refs #4014.
+/// Odds / cargo / ship-loss helpers extracted Refs #4014, #4299.
 
 /// Result of applying trade interception: reduced delivered amounts and fleet updates.
 class TradeInterceptionResult {
@@ -28,17 +30,6 @@ class TradeInterceptionResult {
 
   final Map<CommodityId, int> reducedDelivered;
   final List<Fleet> updatedFleets;
-}
-
-/// Pure odds bag for cargo intercept and merchant ship loss (Refs #4014).
-class TradeInterceptionOdds {
-  const TradeInterceptionOdds({
-    required this.pCargoEffective,
-    required this.pShip,
-  });
-
-  final double pCargoEffective;
-  final double pShip;
 }
 
 void _logExtractionAutoTransportInterception(
@@ -62,111 +53,6 @@ void _logExtractionAutoTransportInterception(
     'deliveredBeforeUnits=$beforeUnits deliveredAfterUnits=$afterUnits '
     'perCommodityDelta=$deltaStr',
   );
-}
-
-/// Computes P_cargo_effective and P_ship from a fleet [scan] (GDD formulas).
-TradeInterceptionOdds computeTradeInterceptionOdds(TradeInterceptionScan scan) {
-  final total = scan.interceptScore + scan.evasionScore;
-  final ratio = total > 0 ? scan.interceptScore / total : 1.0;
-
-  // Escort protection: lossReduction = min(0.5, escortStrength/cargoStrength × 0.3). SPEC.
-  final cargoStrength = scan.playerMerchantShips > 0
-      ? scan.playerMerchantShips.toDouble()
-      : 1.0;
-  final escortFactor =
-      (scan.escortStrength / cargoStrength * escortStrengthWeight).clamp(
-        0.0,
-        escortFactorMax,
-      );
-
-  // Base before escort: used for cargo (with escort) and for ship loss (escort applied once per GDD).
-  double baseBeforeEscort = actionFactorPatrol * ratio * civilianTargetBonus;
-  if (scan.hasBlockade) baseBeforeEscort *= blockadeBonusFactor;
-  baseBeforeEscort = baseBeforeEscort.clamp(0.0, 1.0);
-
-  final base = baseBeforeEscort * (1.0 - escortFactor);
-
-  final pIntercept = (1.2 * base).clamp(0.1, 0.9);
-  final raidEfficiency =
-      raidEfficiencyMin + ratio * (raidEfficiencyMax - raidEfficiencyMin);
-  final pCargoEffective = (pIntercept * raidEfficiency).clamp(0.0, 1.0);
-
-  // GDD: shipLossChance = baseShipLoss × (1 - escortFactor) × civilianPenalty — escort applied once.
-  final baseShipLoss = (0.4 * baseBeforeEscort).clamp(0.0, 1.0);
-  final pShip = (baseShipLoss * (1.0 - escortFactor) * civilianShipLossPenalty)
-      .clamp(0.02, 0.5);
-
-  return TradeInterceptionOdds(
-    pCargoEffective: pCargoEffective,
-    pShip: pShip,
-  );
-}
-
-/// Reduces each commodity quantity by (1 - [pCargoEffective]), dropping zeros.
-Map<CommodityId, int> reduceDeliveredForCargoIntercept(
-  Map<CommodityId, int> overseasDelivered,
-  double pCargoEffective,
-) {
-  final reducedDelivered = <CommodityId, int>{};
-  for (final entry in overseasDelivered.entries) {
-    final qty = entry.value;
-    final keep = (qty * (1.0 - pCargoEffective)).round();
-    if (keep > 0) reducedDelivered[entry.key] = keep;
-  }
-  return reducedDelivered;
-}
-
-/// Deterministic LCG trial: how many of [merchantShipCount] ships are lost.
-int countMerchantShipsLost({
-  required int merchantShipCount,
-  required double pShip,
-  required int seed,
-}) {
-  var rng = seed;
-  int nextInt(int max) {
-    if (max <= 0) return 0;
-    rng =
-        (rng * kDeterministicLcgMultiplierGlibc +
-            kDeterministicLcgIncrementGlibc) &
-        kDeterministicLcg31Mask;
-    return rng % max;
-  }
-
-  var shipsToRemove = 0;
-  for (var i = 0; i < merchantShipCount; i++) {
-    if (nextInt(100) < (pShip * 100)) shipsToRemove++;
-  }
-  return shipsToRemove;
-}
-
-/// Removes up to [shipsToRemove] merchant ships from [playerId]'s fleets.
-List<Fleet> applyMerchantShipLoss({
-  required List<Fleet> fleets,
-  required String playerId,
-  required int shipsToRemove,
-}) {
-  if (shipsToRemove <= 0) {
-    return fleets;
-  }
-  var remainingToRemove = shipsToRemove;
-  final updatedFleets = <Fleet>[];
-  for (final f in fleets) {
-    if (f.ownerId != playerId) {
-      updatedFleets.add(f);
-      continue;
-    }
-    final inst = List<ShipInstance>.from(f.ships);
-    for (var i = inst.length - 1; i >= 0 && remainingToRemove > 0; i--) {
-      if (kMerchantShipTypeIds.contains(inst[i].typeId)) {
-        inst.removeAt(i);
-        remainingToRemove--;
-      }
-    }
-    if (inst.isNotEmpty) {
-      updatedFleets.add(f.copyWith(ships: inst));
-    }
-  }
-  return updatedFleets;
 }
 
 /// Apply trade interception: reduce delivered cargo and optionally remove merchant ships.
