+++ packages/colonizethis_economy/test/economy/resource_extractor_test.dart
